+++ src/components/common/Alert/LogOutStyle.jsx
